requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
black==24.3.0
cached-property==1.5.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Django==4.2.11
django-embed-video==1.4.9
exceptiongroup==1.2.0
ffmpeg-python==0.2.0
future==1.0.0
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.6
importlib-metadata==6.7.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.14.0
opencv-python==4.9.0.80
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.2
pyrebase4==4.8.0
python-dotenv==0.21.1
pytube==15.0.0
pytz==2024.1
requests==2.31.0
rsa==4.9
setuptools==69.5.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.7.1
tzdata==2024.1
uritemplate==4.1.1
urllib3
zipp==3.15.0